refactor: util 함수 이동 및 타입 수정
