+++ apps/client/utils/common.ts
@@ -1,5 +1,4 @@
 import { LOGIN } from '@/constants/constant';
-import { fetchClient } from '@/services';
 
 export function getHeaders() {
   if (typeof sessionStorage === 'undefined') return;
@@ -18,19 +17,6 @@ export function goPageWithSelector(selector: Util.LogInState, router: any) {
   if (typeof url !== 'string') return;
   if (url.includes('need_login')) router.back();
   return router.push(url);
-}
-
-export function doSeeMoreMbtiTests({ fetchOption, data, page }: Util.doSeeMoreMbtiTestsProp) {
-  fetchClient(fetchOption).then((response) => {
-    const oldMbtiTestData = data.mbtiTestDataList?.testCoverDTOList;
-    const newMbtiTestData = oldMbtiTestData ? [...oldMbtiTestData, response?.dataList.testCoverDTOList].flat() : [];
-
-    data.setMbtiTestData((prev: Model.MbtiTest[]) => ({
-      ...prev,
-      dataList: { hasNextPage: response?.dataList.hasNextPage, testCoverDTOList: newMbtiTestData },
-    }));
-    page.setPage(page.page + 1);
-  });
 }
 
 export function formatTimeDifference(dateString: string) {
